+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Release NEXT
 - Provide filtering by required tags for resources and template groups endpoints.
 - Created event type for project name update.
 - Fix OpenStack license stats endpoint.
+- Paginate results for customer users endpoint.
 
 Release 0.93.0
 --------------